+++ index.ts
@@ -13,6 +13,19 @@ const CPMAP = new Map([
     // TODO: It need more codepages!
 ]);
 
+interface StdOut extends NodeJS.WriteStream
+{
+    clearLine():void;
+}
+interface StdIn extends NodeJS.ReadStream
+{
+    setRawMode(rawMode:boolean):void;
+}
+const stdout = <StdOut>process.stdout;
+const stdin = <StdIn>process.stdin;
+
+const children = new Set<Spawn>();
+
 // functions
 function spawn(command:string, args?:string[]):child_process.ChildProcessWithoutNullStreams
 {
@@ -46,30 +59,50 @@ function exec(cmd:string):Promise<string>
 class LineDetector
 {
     private buffer:string = '';
+    private breaked:boolean = false;
+
     constructor(private readonly callback:(line:string)=>void)
     {
     }
 
     add(text:string):void
     {
-        for (;;)
+        for (let i=0;i<text.length;i++)
         {
-            let cmdend = text.indexOf('\n');
-            if (cmdend !== -1)
+            const chr = text.charAt(i);
+            switch (chr)
             {
-                const next = cmdend+1;
-                if (text.charAt(cmdend-1) === '\r') cmdend--;
-                this.buffer += text.substr(0, cmdend);
+            case '\r':
+                this.breaked = true;
                 this.callback(this.buffer);
-                text = text.substr(next);
                 this.buffer = '';
-            }
-            else
-            {
-                this.buffer += text;
+                break;
+            case '\b':
+                this.breaked = false;
+                this.buffer = this.buffer.substr(0, this.buffer.length-1);
+                break;
+            case '\n':
+                if (this.breaked)
+                {
+                    this.breaked = false;
+                }
+                else
+                {
+                    this.callback(this.buffer);
+                    this.buffer = '';
+                }
+                break;
+            default:
+                this.breaked = false;
+                this.buffer += chr;
                 break;
             }
         }
+    }
+
+    getBuffer():string
+    {
+        return this.buffer;
     }
 }
 
@@ -79,10 +112,6 @@ export interface Spawn
     on(event: 'stdout', listener: (message: string) => void): this;
     emit(event: 'stdout', message:string): boolean;
 
-    addListener(event: 'stdin', listener: (message: string) => void): this;
-    on(event: 'stdin', listener: (message: string) => void): this;
-    emit(event: 'stdin', message:string): boolean;
-    
     addListener(event: 'close', listener: () => void): this;
     on(event: 'close', listener: () => void): this;
     emit(event: 'close'): boolean;
@@ -92,9 +121,91 @@ export interface Spawn
     emit(event: 'open'): boolean;
 }
 
+const emptyFunc = ()=>{};
+
+export class StdInListener extends LineDetector
+{
+
+    private restoreRequest:NodeJS.Timeout|null = null;
+
+    private readonly onstdin = (key:string)=>{
+        if (this.restoreRequest)
+        {
+            clearTimeout(this.restoreRequest);
+            this.restoreRequest = null;
+            stdout.write(this.getBuffer());
+        }
+        if ( key === '\u0003' ) {
+            for (const child of children)
+            {
+                child.kill('SIGINT');
+            }
+            process.exit(-1);
+        }
+        else
+        {
+            stdout.write(key);
+            if (key === '\b')
+            {
+                stdout.write(' ');
+                stdout.write('\b');
+            }
+            this.add(key);
+        }
+    };
+
+    constructor(listener:(data:string)=>void)
+    {
+        super(listener);
+        stdin.setRawMode(true);
+        stdin.resume();
+        stdin.setEncoding('utf8');
+        stdin.on('error', emptyFunc);
+        stdin.on('data', this.onstdin);
+        
+        const that = this;
+        const oldlog = console.log;
+        console.log = function(message?:any, ...params:any[]){
+            if (that.getBuffer() !== '')
+            {
+                if (that.restoreRequest)
+                {
+                    clearTimeout(that.restoreRequest);
+                }
+                that.restoreRequest = setTimeout(()=>{
+                    that.restoreRequest = null;
+                    stdout.write(that.getBuffer());
+                }, 100);
+
+                stdout.write('\r');
+                stdout.clearLine();
+                oldlog.apply(this, arguments as any);
+            }
+            else
+            {
+                oldlog.apply(this, arguments as any);
+            }
+        };
+
+    }
+
+    remove():void
+    {
+        stdin.removeListener('error', emptyFunc);
+        stdin.removeListener('data', this.onstdin);
+        stdin.end();
+    }
+}
+
+export interface Removable
+{
+    remove():void;
+}
+
 export class Spawn extends EventEmitter
 {
-    private spawned?:child_process.ChildProcessWithoutNullStreams;
+    private spawned:child_process.ChildProcessWithoutNullStreams|null = null;
+    private killed = false;
 
     stdin(message:string):void
     {
@@ -109,6 +220,8 @@ export class Spawn extends EventEmitter
     constructor(command:string, args?:string[])
     {
         super();
+        children.add(this);
+
         (async()=>{
             const isWindows = os.platform().startsWith('win32');        
             if (isWindows)
@@ -121,35 +234,32 @@ export class Spawn extends EventEmitter
                 command = command.replace(/\//g, '\\');
                 let nargs = ['/s', '/c', command]; // for call global binary
                 if (args) nargs = nargs.concat(args);
+                if (this.killed)
+                {
+                    this.emit('close');
+                    return;
+                }
                 this.spawned = spawn('cmd', nargs);
             }
             else
             {
-                this.spawned = spawn(command, args);
                 await new Promise(resolve=>{ setTimeout(resolve, 0); }); // sync with windows
+                if (this.killed)
+                {
+                    this.emit('close');
+                    return;
+                }
+                this.spawned = spawn(command, args);
             }
 
-            const stdin = new LineDetector(command=>{
-                if (!this.emit('stdin', command))
-                {
-                    this.stdin(command);
-                }
-            });
             const stdout = new LineDetector(out=>{
                 if (!this.emit('stdout', out))
                 {
                     console.log(out);
                 }
             });
-            function onstdin(chunk:Buffer):void
-            {
-                stdin.add(chunk.toString('utf8'));
-            }
-            process.stdin.on('error', ()=>{});
-            process.stdin.on('data', onstdin);
             this.spawned.on('close', ()=>{
-                process.stdin.removeListener('data', onstdin);
-                process.stdin.end();
+                children.delete(this);
                 this.emit('close');
             });
             this.spawned.stdout.on('data', chunk=>{
@@ -162,6 +272,17 @@ export class Spawn extends EventEmitter
             });
             this.emit('open');
         })();
+    }
+
+    kill(signal?:string):void
+    {
+        if (this.killed) return;
+        this.killed = true;
+        if (this.spawned)
+        {
+            this.spawned.kill(signal);
+            this.spawned = null;
+        }
     }
 }